Backport bzlmod support to 25.x
All this change does is apply the BCR patches from
[here](https://github.com/bazelbuild/bazel-central-registry/tree/main/modules/protobuf/25.6/patches),
with the exception of the `rules_ruby` change which has already landed
in 2dac90d6459fd46c874af0bee6c277882eaa329a.

The motivation for this is that on main we still take a dependency on
v25 for the purpose of compatibility testing, and making 25.x compatible
with bzlmod will make it easier to get rid of our remaining
WORKSPACE-only dependencies.